LPS-6391
git-svn-id: svn://svn.liferay.com/repos/public/portal/trunk@42119 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portal-impl/src/com/liferay/portal/upload/UploadServletRequestImpl.java
@@ -57,7 +57,7 @@ import org.apache.commons.fileupload.servlet.ServletFileUpload;
 public class UploadServletRequestImpl
 	extends HttpServletRequestWrapper implements UploadServletRequest {
 
-	public static final int DEFAULT_SIZE_MAX = GetterUtil.getInteger(
+	public static final long DEFAULT_SIZE_MAX = GetterUtil.getLong(
 		PropsUtil.get(UploadServletRequestImpl.class.getName() + ".max.size"));
 
 	public static final File DEFAULT_TEMP_DIR = new File(
